+++ src/int32.ml
@@ -137,7 +137,7 @@ let of_nativeint_trunc = Conv.nativeint_to_int32_trunc
 let to_nativeint = Conv.int32_to_nativeint
 let to_nativeint_exn = to_nativeint
 
-let pow b e = of_int_exn (Int_math.int_pow (to_int_exn b) (to_int_exn e))
+let pow b e = of_int_exn (Int_math.Private.int_pow (to_int_exn b) (to_int_exn e))
 let ( ** ) b e = pow b e
 
 module Pow2 = struct