+++ src/int_math.mli
@@ -1,11 +1,9 @@
-(** This module is not exposed in Core.  Instead, these functions are accessed and
-    commented in the various Core modules implementing [Int_intf.S]. *)
+(** This module implements derived integer operations (e.g., modulo, rounding to
+    multiples) based on other basic operations. *)
 
 open! Import
 
-(*_ This interface is not defined in int_intf.ml because we don't want users of Core to
-  think about it. *)
-module type T = sig
+module type Make_arg = sig
   type t
   include Floatable.S  with type t := t
   include Stringable.S with type t := t
@@ -25,15 +23,25 @@ module type T = sig
   val rem : t -> t -> t
 end
 
-(** derived operations common to various integer modules *)
-module Make (X : T) : sig
+(** Derived operations common to various integer modules.
+
+    See {{!Base.Int.S_common}[Int.S_common]} for a description of the operations derived
+    by this module. *)
+module Make (X : Make_arg) : sig
   val ( %  ) : X.t -> X.t -> X.t
   val ( /% ) : X.t -> X.t -> X.t
   val ( // ) : X.t -> X.t -> float
   include Int_intf.Round with type t := X.t
 end
 
-val int_pow   : int   -> int   -> int
-val int64_pow : int64 -> int64 -> int64
+(*_ See the Jane Street Style Guide for an explanation of [Private] submodules:
 
-val int63_pow_on_int64 : int64 -> int64 -> int64
+  https://opensource.janestreet.com/standards/#private-submodules *)
+module Private : sig
+  val int_pow   : int   -> int   -> int
+  val int64_pow : int64 -> int64 -> int64
+
+  val int63_pow_on_int64 : int64 -> int64 -> int64
+
+  module Pow_overflow_bounds = Pow_overflow_bounds
+end